# openapi.yaml
openapi: 3.1.0
info:
  title: AgriCRM API
  description: Farmer, FPO & Sugar Mill CRM for Theta Analytics
  version: 1.0.0
paths:
...
  /api/v1/invoices/:
    get:
      tags:
      - billing
      summary: Invoice List
      description: "The register.\n\n\U0001F534 An unknown query parameter is a 400,\
        \ not a shrug. FastAPI ignores\nextras by default; `deps.reject_unknown_filters`\
        \ reinstates the rule the\nDjango service enforced by hand, because a filter\
        \ that silently does\nnothing is how someone exports the whole register believing\
        \ they exported\none customer."
      operationId: invoice_list_api_v1_invoices__get
      security:
      - HTTPBearer: []
      parameters:
      - name: entity_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Code
      - name: financial_year
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Financial Year
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: outstanding
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Outstanding
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoicePage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - billing
      summary: Invoice Create
      description: Create a draft. A number is allocated at issue, never here.
      operationId: invoice_create_api_v1_invoices__post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/invoices/extract/:
    post:
      tags:
      - billing
      summary: Invoice Extract
      description: "Read an uploaded document and return a draft the create form can\
        \ load.\n\nEvery attempt is recorded in `crm.invoice_extraction`, including\
        \ the\nfailures \u2014 a model that misread a document is something to look\
        \ at later,\nand a failure that leaves no trace is a failure nobody learns\
        \ from."
      operationId: invoice_extract_api_v1_invoices_extract__post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_invoice_extract_api_v1_invoices_extract__post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Invoice Extract Api V1 Invoices Extract  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/invoices/extract/{extraction_id}/accept/:
    post:
      tags:
      - billing
      summary: Invoice Extract Accept
      description: "Record that a human accepted this extraction onto a draft.\n\n\
        \U0001F534 What was accepted is stored beside what the model proposed. That\
        \ pairing\nis the evaluation set: a golden case built from a real correction\
        \ is worth\nmore than one invented by guessing at what is hard."
      operationId: invoice_extract_accept_api_v1_invoices_extract__extraction_id__accept__post
      security:
      - HTTPBearer: []
      parameters:
      - name: extraction_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Extraction Id
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_invoice_extract_accept_api_v1_invoices_extract__extraction_id__accept__post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Invoice Extract Accept Api V1 Invoices Extract  Extraction
                  Id  Accept  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_invoice_extract_accept_api_v1_invoices_extract__extraction_id__accept__post:
      properties:
        invoice:
          type: string
          format: uuid
          title: Invoice
      type: object
      required:
      - invoice
      title: Body_invoice_extract_accept_api_v1_invoices_extract__extraction_id__accept__post
    Body_invoice_extract_api_v1_invoices_extract__post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
          description: A PDF or a photo of the invoice.
        entity_code:
          type: string
          title: Entity Code
          default: TEPL
        organisation:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Organisation
      type: object
      required:
      - file
      title: Body_invoice_extract_api_v1_invoices_extract__post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InvoiceCreate:
      properties:
        billing_entity:
          type: string
          format: uuid
          title: Billing Entity
        invoice_date:
          type: string
          format: date
          title: Invoice Date
        due_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Due Date
        buyer_name:
          type: string
          title: Buyer Name
        buyer_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Buyer Address
        buyer_gstin:
          anyOf:
          - type: string
          - type: 'null'
          title: Buyer Gstin
        buyer_state_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Buyer State Code
        buyer_is_govt_uin:
          type: boolean
          title: Buyer Is Govt Uin
          default: false
        buyer_order_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Buyer Order No
        work_order_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Work Order Ref
        letter_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Letter Ref
        payment_terms:
          anyOf:
          - type: string
          - type: 'null'
          title: Payment Terms
        organisation:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Organisation
        tax_treatment:
          type: string
          title: Tax Treatment
          default: igst
        tax_rate_pct:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Tax Rate Pct
          default: '18.00'
        place_of_supply_state_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Place Of Supply State Code
        lines:
          items:
            $ref: '#/components/schemas/InvoiceLineIn'
          type: array
          title: Lines
          default: []
      type: object
      required:
      - billing_entity
      - invoice_date
      - buyer_name
      title: InvoiceCreate
    InvoiceDetail:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        invoice_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Invoice No
        entity_code:
          type: string
          title: Entity Code
        invoice_date:
          type: string
          format: date
          title: Invoice Date
        financial_year:
          anyOf:
          - type: string
          - type: 'null'
          title: Financial Year
        buyer_name:
          type: string
          title: Buyer Name
        status:
          type: string
          title: Status
        tax_treatment:
          type: string
          title: Tax Treatment
        taxable_value:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Taxable Value
        tax_amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Tax Amount
        total_value:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Value
        amount_outstanding:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Amount Outstanding
        display:
          $ref: '#/components/schemas/MoneyDisplay'
        buyer_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Buyer Address
        buyer_gstin:
          anyOf:
          - type: string
          - type: 'null'
          title: Buyer Gstin
        buyer_order_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Buyer Order No
        work_order_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Work Order Ref
        letter_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Letter Ref
        payment_terms:
          anyOf:
          - type: string
          - type: 'null'
          title: Payment Terms
        tax_rate_pct:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Tax Rate Pct
        amount_in_words:
          anyOf:
          - type: string
          - type: 'null'
          title: Amount In Words
        amount_received:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Amount Received
        issued_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Issued At
        cancelled_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Cancelled At
        cancellation_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Cancellation Reason
        lines:
          items:
            $ref: '#/components/schemas/InvoiceLineOut'
          type: array
          title: Lines
        payments:
          items:
            $ref: '#/components/schemas/InvoicePaymentOut'
          type: array
          title: Payments
      type: object
      required:
      - id
      - invoice_no
      - entity_code
      - invoice_date
      - financial_year
      - buyer_name
      - status
      - tax_treatment
      - taxable_value
      - tax_amount
      - total_value
      - amount_outstanding
      - display
      - buyer_address
      - buyer_gstin
      - buyer_order_no
      - work_order_ref
      - letter_ref
      - payment_terms
      - tax_rate_pct
      - amount_in_words
      - amount_received
      - issued_at
      - cancelled_at
      - cancellation_reason
      - lines
      - payments
      title: InvoiceDetail
    InvoiceLineIn:
      properties:
        line_no:
          anyOf:
          - type: integer
          - type: 'null'
          title: Line No
        description:
          type: string
          title: Description
        hsn_sac:
          anyOf:
          - type: string
          - type: 'null'
          title: Hsn Sac
        quantity:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Quantity
        unit:
          type: string
          title: Unit
        rate:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Rate
        rate_is_tax_inclusive:
          type: boolean
          title: Rate Is Tax Inclusive
          default: false
        tax_rate_pct:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Tax Rate Pct
        location_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Note
      type: object
      required:
      - description
      - quantity
      - unit
      - rate
      title: InvoiceLineIn
      description: "A line as the client sends it.\n\n\U0001F534 No amounts. `line_taxable_value`,\
        \ `line_tax_amount` and `line_total` are\ncomputed server-side in `api/money.py`\
        \ \u2014 a client that could post its own\ntotals could post an invoice whose\
        \ lines do not sum to its header."
    InvoiceLineOut:
      properties:
        line_no:
          type: integer
          title: Line No
        description:
          type: string
          title: Description
        hsn_sac:
          anyOf:
          - type: string
          - type: 'null'
          title: Hsn Sac
        quantity:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Quantity
        unit:
          type: string
          title: Unit
        quantity_ha:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Quantity Ha
        rate:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Rate
        line_taxable_value:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Line Taxable Value
        line_tax_amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Line Tax Amount
        line_total:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Line Total
      type: object
      required:
      - line_no
      - description
      - hsn_sac
      - quantity
      - unit
      - quantity_ha
      - rate
      - line_taxable_value
      - line_tax_amount
      - line_total
      title: InvoiceLineOut
    InvoicePage:
      properties:
        count:
          type: integer
          title: Count
        results:
          items:
            $ref: '#/components/schemas/InvoiceRow'
          type: array
          title: Results
      type: object
      required:
      - count
      - results
      title: InvoicePage
    InvoicePaymentOut:
      properties:
        amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Amount
        amount_display:
          type: string
          title: Amount Display
        received_on:
          type: string
          format: date
          title: Received On
        mode:
          anyOf:
          - type: string
          - type: 'null'
          title: Mode
        reference:
          anyOf:
          - type: string
          - type: 'null'
          title: Reference
      type: object
      required:
      - amount
      - amount_display
      - received_on
      - mode
      - reference
      title: InvoicePaymentOut
    InvoiceRow:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        invoice_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Invoice No
        entity_code:
          type: string
          title: Entity Code
        invoice_date:
          type: string
          format: date
          title: Invoice Date
        financial_year:
          anyOf:
          - type: string
          - type: 'null'
          title: Financial Year
        buyer_name:
          type: string
          title: Buyer Name
        status:
          type: string
          title: Status
        tax_treatment:
          type: string
          title: Tax Treatment
        taxable_value:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Taxable Value
        tax_amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Tax Amount
        total_value:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Value
        amount_outstanding:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Amount Outstanding
        display:
          $ref: '#/components/schemas/MoneyDisplay'
      type: object
      required:
      - id
      - invoice_no
      - entity_code
      - invoice_date
      - financial_year
      - buyer_name
      - status
      - tax_treatment
      - taxable_value
      - tax_amount
      - total_value
      - amount_outstanding
      - display
      title: InvoiceRow
...
    MoneyDisplay:
      properties:
        total:
          type: string
          title: Total
        outstanding:
          type: string
          title: Outstanding
        taxable:
          type: string
          title: Taxable
        tax:
          type: string
          title: Tax
        received:
          type: string
          title: Received
      type: object
      required:
      - total
      - outstanding
      - taxable
      - tax
      - received
      title: MoneyDisplay
      description: "Pre-formatted figures. See `backend/money.py` for why the server\
        \ does this.\n\n\U0001F534 Indian digit grouping is done here and never in\
        \ TypeScript. The detail\nscreen reads `taxable` and `tax` as well as `total`\
        \ and `outstanding`; they\nwere absent, so it rendered \"\u20B9undefined\"\
        \ for two of the four figures on\nevery invoice."
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer